Add forked AddPathPlugin and removed required patch
README.md contained a reference to an outdated required patch (103)
for php-http/client-common:1.x. The composer.json contained the correct
reference (113). Although, that patch also became outdated because it
did not contain a bulletproof fix for the problem.
The issue has been properly fixed in the 2.x branch. This is the reason
why we forked the fixed version from the 2.x branch and with that
also removed all the overhead caused by conditional patching the
php-http/client-common dependency.

https://github.com/php-http/client-common/issues/171